Started on $1 recognizer
Added a lot of stuff, but I have yet to test it. My plan is to lay out a template and then test and debug components one after the other. I believe that this will be easier and faster than the alternative, which would be to test them individually.

I feel that once $1 is complete then it will be easier to complete the other 2. This is mainly because I am still working on familiarizing myself with Numpy's idiosyncrasies. An endeavor I believe to be very worthwhile. I also have plenty of time. I may not be able to clean up the UI as neatly as I would like, but eveything else is going well